refactor(library-config): reorganize Linux process context
