+++ libdd-library-config/src/lib.rs
@@ -2,6 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 extern crate alloc;
 
+#[cfg(all(
+    target_os = "linux",
+    any(feature = "process-context-reader", feature = "process-context-writer")
+))]
 pub mod otel_process_ctx;
 pub mod tracer_metadata;
